Moved some functions to a more appropriate file
Left those out by accident.
